drivers: syscon: use uint32_t for register offsets
Use uint32_t to allow larger offsets in place of the existing uint16_t
configuration.

+++ include/zephyr/drivers/syscon.h
@@ -44,13 +44,13 @@ typedef int (*syscon_api_get_base)(const struct device *dev, uintptr_t *addr);
  * @brief Read a single register.
  * See syscon_read_reg() for argument description.
  */
-typedef int (*syscon_api_read_reg)(const struct device *dev, uint16_t reg, uint32_t *val);
+typedef int (*syscon_api_read_reg)(const struct device *dev, uint32_t reg, uint32_t *val);
 
 /**
  * @brief Write a single register.
  * See syscon_write_reg() for argument description.
  */
-typedef int (*syscon_api_write_reg)(const struct device *dev, uint16_t reg, uint32_t val);
+typedef int (*syscon_api_write_reg)(const struct device *dev, uint32_t reg, uint32_t val);
 
 /**
  * @brief Atomically update bits in a register.
@@ -118,9 +118,9 @@ static inline int z_impl_syscon_get_base(const struct device *dev, uintptr_t *ad
  * @retval 0 on success.
  * @retval -ENOSYS If the API or function isn't implemented.
  */
-__syscall int syscon_read_reg(const struct device *dev, uint16_t reg, uint32_t *val);
+__syscall int syscon_read_reg(const struct device *dev, uint32_t reg, uint32_t *val);
 
-static inline int z_impl_syscon_read_reg(const struct device *dev, uint16_t reg, uint32_t *val)
+static inline int z_impl_syscon_read_reg(const struct device *dev, uint32_t reg, uint32_t *val)
 {
 	const struct syscon_driver_api *api = DEVICE_API_GET(syscon, dev);
 
@@ -144,9 +144,9 @@ static inline int z_impl_syscon_read_reg(const struct device *dev, uint16_t reg,
  * @retval 0 on success.
  * @retval -ENOSYS If the API or function isn't implemented.
  */
-__syscall int syscon_write_reg(const struct device *dev, uint16_t reg, uint32_t val);
+__syscall int syscon_write_reg(const struct device *dev, uint32_t reg, uint32_t val);
 
-static inline int z_impl_syscon_write_reg(const struct device *dev, uint16_t reg, uint32_t val)
+static inline int z_impl_syscon_write_reg(const struct device *dev, uint32_t reg, uint32_t val)
 {
 	const struct syscon_driver_api *api = DEVICE_API_GET(syscon, dev);
